requirements.txt
pandas
numpy
pymongo
pyspark
scikit-learn
matplotlib
flask
flask-cors
joblib
python-dotenv
